A facade object to use Shiny Progress with dplyr

Used to translate dplyr progress to a Shiny progress bar

Public fields

proxy

Shiny Progress object

n

total ticks for dplyr progress; can be reinitialized

i

current tick for dplyr progress; can be reinitialized

offset

offset for Shiny Progress

amount

total advanced amount in Shiny after the dplyr progress is done

Methods

Public methods

ShinyToDplyrProgressFacade$new()

ShinyToDplyrProgressFacade$tick()

ShinyToDplyrProgressFacade$stop()

ShinyToDplyrProgressFacade$print()

ShinyToDplyrProgressFacade$clone()

Method new()

Initialize a shiny progress to dplyr facade. This can either be called with the `proxy` and `amount` arguments for a first initialization (e.g. when using `new`), or with `n` and `min_time` arguments for the dplyr-like initialization

Usage

ShinyToDplyrProgressFacade$new(
  n = 1,
  min_time = 0,
  proxy = NULL,
  amount = 1,
  ...
)

Arguments

n

Total number of items

min_time

Progress bar will wait until at least `min_time` seconds have elapsed before displaying any results.

proxy

Shiny Progress object

amount

amount advanced in Shiny Progress

...

ignored

Returns

an object that can be used in `.progress` arguments for dplyr

Method tick()

Advance a tick

Usage

ShinyToDplyrProgressFacade$tick()

Method stop()

Stop the timer; advance until the end

Usage

ShinyToDplyrProgressFacade$stop()

Method print()

Print out the progress; transfers the progress to Shiny

Usage

ShinyToDplyrProgressFacade$print(...)

Arguments

...

ignored

Method clone()

The objects of this class are cloneable with this method.

Usage

ShinyToDplyrProgressFacade$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.