MPC estimate method.

MPC estimate method.

# S3 method for tdmore_mpc
estimate(
  object,
  observed = NULL,
  regimen = NULL,
  covariates = NULL,
  par = NULL,
  fix = NULL,
  method = "L-BFGS-B",
  se.fit = TRUE,
  lower = NULL,
  upper = NULL,
  multistart = F,
  control = list(),
  ...,
  .progress = NULL
)

Arguments

object
a tdmore, tdmore_set or tdmore_mixture object

observed
data frame with at least a TIME column, and all observed data. The observed data will be compared to the model predictions. If not specified, we estimate the population prediction

regimen
data frame describing the treatment regimen.

covariates
the model covariates

par
optional starting parameter for the MLE minimization

fix
named numeric vector with the fixed parameters. Names can repeat to fix parameters in multiple occasions.

method
the optimisation method, by default, method "L-BFGS-B" is used Can also be specified as a list, in which case all methods specified will be tried and the one finding the estimate with highest log-likelihood will be returned.

se.fit
calculate the variance-covariance matrix for the fit Putting this to FALSE can reduce computation time.

lower
the lower bounds of the parameters, if null, -5 * sqrt(diag(model$omega)) is used

upper
the upper bounds of the parameters, if null, +5 * sqrt(diag(model$omega)) is used

multistart
if TRUE, perform optimization using many different starting conditions. A combination of -sd, +sd for each non-IOV parameter is explored. This avoids returning a local optimum, and strengthens the belief that the found optimum is a global optimum.

Alternatively, a named vector with the requested perturbations to parameters can be provided: estimate(multistart=c(eta_Tlag=0.5)). This will evaluate the provided starting value for eta_Tlag +- 0.5, without also perturbing the other initial values.

WARNING: This may lead to extremely long computation times. It is recommended to use a control argument to reduce accuracy and report estimation progress.

control
control options for `estimate` If trace > 0, the estimation function also reports errors when calculating log-likelihood.

...
extra parameters to pass to the optim function A good example is to specify `control=list(trace=1, REPORT=10, factr=1e13)` to improve performance. This will `trace` the estimation progress and `report` every 10 iterations. It limits the precision of the `L-BFGS-B` method to 3 significant digits (`factr` is multiplied with the machine precision, `1e12 * 1e-16 = 1e-4`).

Incidentally, 3 significant digits is the default value for NONMEM. Without the above option, `estimate` will calculate with a relative precision of `1e-8`.

.progress
progress bar definition

Value

a tdmorefit