Find the dose required (quick and dirty, to be optimised)

Find the dose required (quick and dirty, to be optimised)

findDose(tdmorefit, regimen, doseRows = NULL, interval = c(0, 1e+10),
  target, se.fit = FALSE, level = 0.95, mc.maxpts = 100,
  .progress = "none", .parallel = FALSE, ...)

Arguments

tdmorefit
the tdmorefit object

regimen
the regimen

doseRows
which rows of the regimen to adapt when searching for a new dose, or NULL to take the last one

interval
which interval to search a dose in. Defaults to a ridiculously high range

target
target value, as a data.frame

se.fit
TRUE to provide a confidence interval on the dose prediction, adding columns dose.median, dose.lower and dose.upper

level
the confidence interval on the dose, only used if se.fit is true

mc.maxpts
maximum number of points to sample in Monte Carlo simulation

.progress
see plyr::ddply

.parallel
see plyr::ddply

...
extra arguments passed to stats::uniroot

Value

a recommendation object