Algebraic equations for PK models

Algebraic equations for PK models

pk1cptiv_(t, A1 = 0, TIME, AMT, V, K, SS = 0, II = Inf)

pk1cptinfusion_(t, A1 = 0, TIME, AMT, RATE, V, K, SS = 0, II = Inf)

pk1cptoral_(t, A0 = 0, A1 = 0, TIME, AMT, V, K, KA, SS = 0, II = Inf)

pk2cptiv_(t, A1 = 0, A2 = 0, TIME, AMT, V, K, K12, K21, SS = 0, II = Inf)

pk2cptinfusion_(
  t,
  A1 = 0,
  A2 = 0,
  TIME,
  AMT,
  RATE,
  V,
  K,
  K12,
  K21,
  SS = 0,
  II = Inf
)

pk2cptoral_(
  t,
  A0 = 0,
  A1 = 0,
  A2 = 0,
  TIME,
  AMT,
  V,
  K,
  KA,
  K12,
  K21,
  SS = 0,
  II = Inf
)

pk3cptiv_(
  t,
  A1 = 0,
  A2 = 0,
  A3 = 0,
  TIME,
  AMT,
  V,
  K,
  K12,
  K21,
  K13,
  K31,
  SS = 0,
  II = Inf
)

pk3cptinfusion_(
  t,
  A1 = 0,
  A2 = 0,
  A3 = 0,
  TIME,
  AMT,
  RATE,
  V,
  K,
  K12,
  K21,
  K13,
  K31,
  SS = 0,
  II = Inf
)

pk3cptoral_(
  t,
  A0 = 0,
  A1 = 0,
  A2 = 0,
  A3 = 0,
  TIME,
  AMT,
  V,
  K,
  KA,
  K12,
  K21,
  K13,
  K31,
  SS = 0,
  II = Inf
)

Arguments

t
sampling times

A1
initial state for A1 at TIME

TIME
dosing time

AMT
dosing amount

V
volume of distribution

K
elimination constant

SS
0 for no steady-state, 1 for steady-state

II
inter-dose interval, required for steady-state calculations

RATE
infusion rate

A0
initial state for A0 at TIME

KA
absorption constant

A2
initial state for A2 at TIME

K12
transfer rate from central compartment to peripheral compartment A

K21
transfer rate from peripheral compartment A to central compartment

A3
initial state for A3 at TIME

K13
transfer rate from central compartment to peripheral compartment B

K31
transfer rate from peripheral compartment B to central compartment

Value

vector of same size as `t`, with the concentrations