Plot a tdmore object (typical profile of the population and between subject variability).

Plot a tdmore object (typical profile of the population and between subject variability).

# S3 method for tdmore
plot(x, regimen, covariates = NULL, newdata = NULL,
  bsv = TRUE, mc.subjects = 100, .progress = "none", ...)

Arguments

x
the tdmore object

regimen
the regimen to be predicted

covariates
covariates

newdata
a data.frame with at least TIME and any other columns to plot, NULL to plot all columns from the original observed data between time 0 and max(observationTime) or a numeric vector of times

bsv
add between subject variability, enabled by default

mc.subjects
number of subjects in the monte carlo simulation

.progress
either "none" or "text" to see calculation progress of monte carlo simulations

...
ignored

Value

a ggplot object with the the population prediction