Get the residuals for a specific tdmorefit

Get the residuals for a specific tdmorefit

Get the residual values of a predicted value vs the observed value

# S3 method for tdmorefit
residuals(object, data, weighted = FALSE, ...)

# S3 method for tdmore
residuals(object, predicted, observed, weighted = FALSE,
  inverse = FALSE, ...)

Arguments

object
A TDMore object

data
optional data.frame If specified, this adds the original residuals back onto the data.frame If points at a time without corresponding IRES are present, an IRES of 0 is used and a warning is emitted.

weighted
Should the residuals be divided by the standard deviation of the error model?

...
ignored

predicted
What was predicted, as data.frame

observed
The observed values, as data.frame

inverse
if TRUE, perform the inverse operation (assume `predicted` has IPRED values, and `observed` has IRES/IWRES values)

Value

The `predicted` data.frame, with all output values replaced by IRES or IWRES

Examples

observed <- data.frame(TIME=5, CONC=0.060)
fit <- estimate(tdmore(default_model),
    regimen=data.frame(TIME=0, AMT=10),
    observed=observed,
    covariates=c(WT=70)
    )
wres <- residuals(fit, weighted=TRUE)

all.equal(
  residuals(fit, predict(fit), weighted=TRUE),
  observed
)

#> [1] TRUE

regimen$AMT <- regimen$AMT*2

#> Error in eval(expr, envir, enclos): object 'regimen' not found

predictionForDoubleDose <- residuals(fit, predict(fit, regimen=regimen), weighted=TRUE)

#> Error in predict.tdmorefit(fit, regimen = regimen): object 'regimen' not found