Create a prediction.

Create a prediction.

stat_predict(mapping = NULL, data = NULL, geom = "line",
  position = "identity", color = "steelblue2", ..., newdata = NULL,
  regimen = NULL, parameters = NULL, covariates = NULL,
  show.legend = NA, inherit.aes = TRUE)

Arguments

mapping
Set of aesthetic mappings created by aes() or aes_(). If specified and inherit.aes = TRUE (the default), it is combined with the default mapping at the top level of the plot. You must supply mapping if there is no plot mapping.

data
A tdmorefit object, or NULL to use the tdmorefit object specified in `ggplot()``

geom
The geometric object to use display the data

position
Position adjustment, either as a string, or the result of a call to a position adjustment function.

...
Other arguments passed on to layer(). These are often aesthetics, used to set an aesthetic to a fixed value, like colour = "red" or size = 3. They may also be parameters to the paired geom/stat.

newdata
A data.frame with new data and the columns to predict, or a numeric vector to specify times, and predict all model output or NULL to interpolate between 0 and the maximum known times

regimen
Treatment regimen

parameters
Set parameters. If missing, or if only part of the parameters are specified, the other parameters are taken from the tdmorefit object

covariates
the model covariates, named vector, or data.frame with column 'TIME', and at least TIME 0

show.legend
logical. Should this layer be included in the legends? NA, the default, includes if any aesthetics are mapped. FALSE never includes, and TRUE always includes. It can also be a named logical vector to finely select the aesthetics to display.

inherit.aes
If FALSE, overrides the default aesthetics, rather than combining with them. This is most useful for helper functions that define both data and aesthetics and shouldn't inherit behaviour from the default plot specification, e.g. borders().