Create a TDMore mixture model, based on several options for a discrete covariate.

Create a TDMore mixture model, based on several options for a discrete covariate.

tdmore_mixture_covariates(model, probs, covariates)

Arguments

model
a tdmore model with a discrete covariate

probs
a numeric vector with the a priori probabilities for the covariate values

covariates
a list with covariate values corresponding to the right names

Value

a tdmore_mixture object

Examples

tdmore_mixture_covariates(tacrolimus_storset,
   probs=c(0.5, 0.5),
   covariates=list(c(CYP3A5=1), c(CYP3A5=0))
)

#> $models
#> $models[[1]]
#> Pre-filled covariates: 
#> CYP3A5 
#>      1 
#> Structural model: RxODE 
#> Parameters: ECL EV1 EQ EF EF_IOV EKa_IOV 
#> Covariates: WT HT FEMALE CYP3A5 PredDose FirstDay HCT 
#> Output(s):
#> Cwb  :  ( prop= 0.149  )
#> 
#> $models[[2]]
#> Pre-filled covariates: 
#> CYP3A5 
#>      0 
#> Structural model: RxODE 
#> Parameters: ECL EV1 EQ EF EF_IOV EKa_IOV 
#> Covariates: WT HT FEMALE CYP3A5 PredDose FirstDay HCT 
#> Output(s):
#> Cwb  :  ( prop= 0.149  )
#> 
#> 
#> $probs
#> [1] 0.5 0.5
#> 
#> attr(,"class")
#> [1] "tdmore_mixture"